Desafío_2/requirements.txt
pandas>=2.1.0
python-dotenv>=1.0.0
influxdb-client>=1.37.0